refactor: move register and login page to root route
