+++ src/lib/ucan/capabilities/root.ts
@@ -1,0 +1,8 @@
+import type { Capability } from '@ucans/ucans';
+
+export function rootCapability(): Capability {
+	return {
+		with: { scheme: 'api', hierPart: '*' },
+		can: { namespace: '*', segments: ['*'] }
+	};
+}
