--- src/routes/admin/login/+layout.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	let { children } = $props();
-</script>
-
-<svelte:head>
-	<title>Login | Murmurations Collaborative Cluster Builder</title>
-</svelte:head>
-
-{@render children()}
