--- src/routes/admin/register/+layout.svelte
@@ -1,9 +1,0 @@
-<script lang="ts">
-	let { children } = $props();
-</script>
-
-<svelte:head>
-	<title>Register | Murmurations Collaborative Cluster Builder</title>
-</svelte:head>
-
-{@render children()}
