+++ src/routes/register/+page.ts
